requirements.txt
datasets
openai
together
requests
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
jupyter
python-dotenv
tqdm
